correct returned value type

+++ include/compute_gpu.cuh
@@ -36,7 +36,7 @@ extern float*h_a;
 
 __global__ void compute_kernel(float *d_a, size_t n);
 
-double compute_on_gpu(float*d_a, cudaStream_t stream, int grid, int block, size_t n, double latency_us);
+float compute_on_gpu(float*d_a, cudaStream_t stream, int grid, int block, size_t n, double latency_us);
 void init_vector(int n);
 void free_vector(void);
 #endif
